pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ocean-drift-diffusion"
version = "0.1.0"
description = "Lightweight spatiotemporal diffusion model for probabilistic ocean drift forecasting"
authors = [{name="Yash Thakur"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "torch>=2.2",
    "numpy>=1.24",
    "pandas>=2.0",
    "xarray>=2023.1.0",
    "netCDF4>=1.6.4",
    "tqdm>=4.66.0",
    "matplotlib>=3.7",
    "einops>=0.7.0",
    "pyyaml>=6.0",
    "scipy>=1.10",
]

[project.urls]
homepage = "https://github.com/yourname/ocean-drift-diffusion"
repository = "https://github.com/yourname/ocean-drift-diffusion"

[tool.setuptools.packages.find]
where = ["src"]
